Fix: SearchResult 컴포넌트 분리, 스타일링 적용
1. UploadInfoModal/SearchResult 폴더 생성.
2. 각 장소마다 스타일링 적용
3. 도로명, 지번 둘 다 보이도록 수정

관련 issue: #90
