+++ frontend/src/components/Modal/PostCreateModal/UploadInfoModal/index.tsx
@@ -4,6 +4,7 @@ import styled from "styled-components";
 import COLOR from "@styles/Color";
 import { useDispatch } from "react-redux";
 import Carousel from "@components/Modal/PostCreateModal/UploadInfoModal/Carousel";
+import SearchResult from "@components/Modal/PostCreateModal/UploadInfoModal/SearchResult";
 
 interface FileObject {
   file: File;
@@ -21,20 +22,6 @@ interface IData {
 interface IPagination {
   [key: string]: string;
 }
-
-const SearchResult = ({ searchResult }: any) => {
-  console.log("in SearchResult, searchResult : ", searchResult);
-  return (
-    <SearchResultWrapper>
-      {searchResult.map((location: IData) => (
-        <PlaceWrapper>
-          <PlaceName>{location.place_name}</PlaceName>
-          <AddressName>{location.address_name}</AddressName>
-        </PlaceWrapper>
-      ))}
-    </SearchResultWrapper>
-  );
-};
 
 const UploadInfoModal = ({ changeMode, files }: UploadInfoModalProps) => {
   const [title, setTitle] = useState<string>("");
@@ -100,14 +87,6 @@ const UploadInfoModal = ({ changeMode, files }: UploadInfoModalProps) => {
         alert("검색 결과 중 오류가 발생했습니다.");
         return;
       }
-
-      console.log(`data : ${data}, status : ${status}, pagination : ${pagination}`);
-
-      for (const location of data) {
-        console.log("location : ", location);
-      }
-      console.log("data.length : ", data.length);
-      console.log(pagination);
 
       setSearchResult(data);
     };
@@ -416,27 +395,4 @@ const SearchInput = styled.input`
   &:focus-visible {
     outline: none;
   }
-`;
-
-const SearchResultWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  overflow-y: scroll;
-`;
-
-const PlaceWrapper = styled.div`
-  cursor: pointer;
-  border-bottom: 1px solid ${COLOR.BLACK};
-  &:hover {
-    background: ${COLOR.GRAY};
-  }
-`;
-
-const PlaceName = styled.div`
-  font-size: 1rem;
-  font-weight: bold;
-`;
-
-const AddressName = styled.div`
-  font-size: 0.8rem;
 `;
